Rename type_identifier to type-identifier
PiperOrigin-RevId: 326824775
