+++ validator/cpp/engine/type-identifier.cc
@@ -14,7 +14,7 @@
 // limitations under the license.
 //
 
-#include "type_identifier.h"
+#include "type-identifier.h"
 
 #include <algorithm>
